+++ BTCPayServer.Plugins.PodServer/Services/Podcasts/EpisodesQuery.cs
@@ -5,7 +5,8 @@ public class EpisodesQuery
     public string PodcastId { get; set; }
     public string EpisodeId { get; set; }
     public string SeasonId { get; set; }
-    public string Slug { get; set; }
+    public string PodcastSlug { get; set; }
+    public string EpisodeSlug { get; set; }
     public string ImportGuid { get; set; }
     public bool IncludePodcast { get; set; }
     public bool IncludePeople { get; set; }