+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.3.0]
 
+### Fixed
+
+- Fix public page routes. (#4)
+
 ### Changed
 
 - UI updates.